DBP-382 Rename metrics to metrics source / metrics exporter
