+++ core/scaler.go
@@ -2,7 +2,7 @@ package core
 
 import (
 	"fmt"
-	"scaler/metrics"
+	"scaler/metricssource"
 	"scaler/providers"
 	"scaler/services"
 	s "scaler/shared"
@@ -12,7 +12,7 @@ type ScalerApp struct {
 	appDefinition *s.AppDefinition
 	service       *s.Service
 	provider      *s.Provider
-	metrics       *s.Metrics
+	metricsSource *s.MetricsSource
 }
 
 // TODO: make these configurable
@@ -42,7 +42,7 @@ func InitApp(configPath string) (*ScalerApp, error) {
 		return nil, fmt.Errorf("error while initializing provider: %s", err)
 	}
 
-	metrics, err := initMetrics(&app.MetricsType, configFile)
+	metricsSource, err := initMetricsSource(&app.MetricsSourceType, configFile)
 	if err != nil {
 		return nil, fmt.Errorf("error while initializing metrics: %s", err)
 	}
@@ -51,7 +51,7 @@ func InitApp(configPath string) (*ScalerApp, error) {
 		appDefinition: app,
 		service:       service,
 		provider:      provider,
-		metrics:       metrics,
+		metricsSource: metricsSource,
 	}, nil
 }
 
@@ -98,17 +98,17 @@ func initProvider(t *s.ProviderType, configFile []byte) (*s.Provider, error) {
 	return nil, fmt.Errorf("unknown provider type: %s", *t)
 }
 
-func initMetrics(t *s.MetricsType, configFile []byte) (*s.Metrics, error) {
+func initMetricsSource(t *s.MetricsSourceType, configFile []byte) (*s.MetricsSource, error) {
 	switch *t {
 	case s.Prometheus:
-		prometheus, err := s.LoadConfig[metrics.Prometheus](configFile)
+		prometheus, err := s.LoadConfig[metricssource.Prometheus](configFile)
 		if err != nil {
 			return nil, fmt.Errorf("error while loading prometheus config: %s", err)
 		}
 		if err := prometheus.Init(); err != nil {
 			return nil, fmt.Errorf("error while initializing prometheus: %s", err)
 		}
-		metrics := s.Metrics(prometheus)
+		metrics := s.MetricsSource(prometheus)
 		return &metrics, nil
 	}
 	return nil, fmt.Errorf("unknown metrics type: %s", *t)
