+++ main.go
@@ -2,10 +2,7 @@ package main
 
 import (
 	"flag"
-	"net/http"
 	c "scaler/core"
-
-	"github.com/prometheus/client_golang/prometheus/promhttp"
 )
 
 func main() {
@@ -16,9 +13,5 @@ func main() {
 	if err != nil {
 		panic(err)
 	}
-	go app.Scale()
-
-	http.Handle("/metrics", promhttp.Handler())
-	// TODO make port configurable
-	http.ListenAndServe(":8080", nil)
+	app.Scale()
 }
