+++ metricssource/prometheus.go
@@ -1,4 +1,4 @@
-package metrics
+package metricssource
 
 import (
 	"context"
